requirements.txt
pytest==8.3.5
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-exporter-otlp==1.33.1
opentelemetry-instrumentation-grpc==0.54b1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
pymilvus==2.5.9